Update email.php
Move 

`
            if ($record_entries) {

                Entries::post($newEntry);
            }
`
outside of the $send_email === true if statement. As it stands, you cannot send the output of the form to an email address other than admin _and_ save the entry to the database. Moving this $record_entries if statement outside allows it to be saved as long as the Record Entries tag is added to the action of the form in the settings of the block.